Add RSVP section; remove linting
Add real info to the RSVP section.  Both RSVP and cabin reservation
(payment) are handled by 3rd-party services, so this just consists of
adding links/calls-to-action to these services.  I still need to add
information about the cabin reservation to the 'Accommodation' section.

Minor:
Remove lint step from the build; this has always failed for as long as
I've been working on the project and resulted in totally unnecessary dev
dependencies that made `npm shrinkwrap` fail.  This hasn't been as much
of an issue as I've switched to using `yarn` for install, but still some
value.