Collapse navbar when menu item selected
For a while I had noticed that it was annoying that selection of a menu
item didn't automatically collapse the menu when viewed in a mobile
browser.
Following
https://stackoverflow.com/questions/14203279/bootstrap-close-responsive-menu-on-click#21744007
there is a data attribute one can use to cause clicks to collapse the
bar, but it causes flickering, so two anchor elements need to be
present.  Refactored MainNavLink from a simple anchor to a list item
containing two links: one which displays in small mode and one which
displays in large mode.